refactor(optimizer): résidus MSE/AutoEQ en float64, plus d'arrondi f32 interne
Les résidus minimisés (mesuré − cible) étaient stockés en Float32Array puis
relus par les calculs en aval — un arrondi float32 intermédiaire évitable sur
la grandeur même que l'optimiseur minimise. Passés en Float64Array, purement
internes (aucune frontière REW, aucun cast de parité) :
- optimization/mseGrid.js : deltas
- optimization/frequencyGrid.js : deltas + decDeltas (copie décimée)
- autoeq/SpanAnalyzer.js : splDeltas (relu pour interpolation + notches)

Principe : entrée float32 (REW) → calcul float64 → quantification float32 aux
seules frontières d'export. freqs/weights (entrées) et toutes les frontières
REW/WebAudio/parité restent inchangés.

Goldens invariants : test:all identique au bit d'affichage (Ex1 0.22 / Ex2 5.33
/ Ex3 0.31 / Ex4 2.32), multi-sub-optimizer 35 + 15/15, scorer, characterization,
genetic-algorithm, 512 unit, build — tous verts. L'écart f32→f64 est sous le
quantum d'export, donc les masters ne bougent pas.

+++ src/autoeq/SpanAnalyzer.js
@@ -45,7 +45,7 @@ export class SpanAnalyzer {
 
     // Grille 1/96 PPO (remplie par init)
     this.freqs = null; // Float32Array des fréquences
-    this.splDeltas = null; // Float32Array: measured - target (résiduel initial)
+    this.splDeltas = null; // Float64Array: measured - target (résiduel initial)
     this.numPoints = 0;
   }
 
@@ -68,7 +68,9 @@ export class SpanAnalyzer {
     }
 
     this.freqs = Float32Array.from(freqs);
-    this.splDeltas = new Float32Array(freqs.length);
+    // Résidu (mesuré − cible) relu pour interpolation et détection de notches :
+    // gardé en float64, aucun arrondi f32 intermédiaire.
+    this.splDeltas = new Float64Array(freqs.length);
     this.numPoints = freqs.length;
 
     for (let i = 0; i < this.numPoints; i++) {
